requirements.txt
streamlit
scikit-learn
pandas
numpy
PyMuPDF
python-docx
joblib
lxml
sentence-transformers
spacy
